Extract GetMeterUnitsInPackets template function
